Move consensus state and logic out of `Node`
Instead, we add a new `Consensus` struct which stores the state
and logic and `Node` contains a `Consensus`.

Originally, I was hoping to keep the complexity of `Consensus`
down by only including the direct consensus protocol and omitting
state and transaction execution.

However, having experiemented a bit, I feel like these elements
are too heavily integrated with consensus to make their separation
feasible. Therefore, `Consensus` also holds the `State` and is
responsible for calling `State::apply_transaction`, etc.

I have further plans to refactor the consensus, but I think it
makes sense to merge this first and raise further PRs for
subsequent changes.
