+++ zilliqa/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod api;
 pub mod cfg;
+pub mod consensus;
 pub mod crypto;
 mod exec;
 pub mod message;
